Move default OpenAI model into typechat proper.
We can figure out finer details later.
