+++ python/src/typechat/__init__.py
@@ -2,12 +2,13 @@
 #
 # SPDX-License-Identifier: MIT
 
-from typechat._internal.model import TypeChatModel
+from typechat._internal.model import DefaultOpenAIModel, TypeChatModel
 from typechat._internal.result import Failure, Result, Success
 from typechat._internal.translator import TypeChatTranslator
 from typechat._internal.validator import TypeChatValidator
 
 __all__ = [
+    "DefaultOpenAIModel",
     "TypeChatModel",
     "TypeChatTranslator",
     "TypeChatValidator",
